+++ ui/component/publishFormErrors/view.jsx
@@ -46,10 +46,11 @@ function PublishFormErrors(props: Props) {
       {!bid && <div>{__('A deposit amount is required')}</div>}
       {bidError && <div>{__('Please check your deposit amount.')}</div>}
       {isUploadingThumbnail && <div>{__('Please wait for thumbnail to finish uploading')}</div>}
-      {!isUploadingThumbnail && !thumbnail && (
+      {!isUploadingThumbnail && !thumbnail ? (
         <div>{__('A thumbnail is required. Please upload or provide an image URL above.')}</div>
+      ) : (
+        thumbnailError && !thumbnailUploaded && <div>{__('Thumbnail is invalid.')}</div>
       )}
-      {thumbnailError && !thumbnailUploaded && <div>{__('Thumbnail is invalid.')}</div>}
       {editingURI && !isStillEditing && !filePath && (
         <div>{__('Please reselect a file after changing the LBRY URL')}</div>
       )}