+++ ui/component/selectThumbnail/index.js
@@ -6,23 +6,19 @@ import {
   selectFileInfosByOutpoint,
   doResetThumbnailStatus,
 } from 'lbry-redux';
-
 import { doOpenModal } from 'redux/actions/app';
 import PublishPage from './view';
 
-const select = state => ({
+const select = (state) => ({
   ...selectPublishFormValues(state),
   fileInfos: selectFileInfosByOutpoint(state),
   myClaimForUri: selectMyClaimForUri(state),
 });
 
-const perform = dispatch => ({
-  updatePublishForm: value => dispatch(doUpdatePublishForm(value)),
+const perform = (dispatch) => ({
+  updatePublishForm: (value) => dispatch(doUpdatePublishForm(value)),
   resetThumbnailStatus: () => dispatch(doResetThumbnailStatus()),
   openModal: (modal, props) => dispatch(doOpenModal(modal, props)),
 });
 
-export default connect(
-  select,
-  perform
-)(PublishPage);
+export default connect(select, perform)(PublishPage);