+++ ui/modal/modalConfirmThumbnailUpload/index.js
@@ -3,13 +3,10 @@ import { doHideModal } from 'redux/actions/app';
 import { doUploadThumbnail, doUpdatePublishForm } from 'lbry-redux';
 import ModalConfirmThumbnailUpload from './view';
 
-const perform = dispatch => ({
+const perform = (dispatch) => ({
   closeModal: () => dispatch(doHideModal()),
-  upload: file => dispatch(doUploadThumbnail(null, file, null, null, file.path)),
-  updatePublishForm: value => dispatch(doUpdatePublishForm(value)),
+  upload: (file, cb) => dispatch(doUploadThumbnail(null, file, null, null, file.path, cb)),
+  updatePublishForm: (value) => dispatch(doUpdatePublishForm(value)),
 });
 
-export default connect(
-  null,
-  perform
-)(ModalConfirmThumbnailUpload);
+export default connect(null, perform)(ModalConfirmThumbnailUpload);